feat(l10n): add applyTransformedLocalizedStrings utility function
