+++ .changeset/odd-years-think.md
@@ -1,0 +1,6 @@
+---
+'@commercetools-frontend/l10n': minor
+'playground': minor
+---
+
+Expose utility function `applyTransformedLocalizedStrings`.
